entropy.c: explicit conversion to size_t
lenread can be -1, 0, or a value up to the number of bytes requested.  We
already handled the case of -1 and 0, so it is safe to cast to size_t.

+++ libcperciva/util/entropy.c
@@ -54,8 +54,8 @@ entropy_read(uint8_t * buf, size_t buflen)
 		}
 
 		/* We've filled a portion of the buffer. */
-		buf += lenread;
-		buflen -= lenread;
+		buf += (size_t)lenread;
+		buflen -= (size_t)lenread;
 	}
 
 	/* Close the device. */
